requirements.txt
absl-py==1.4.0
blinker==1.4
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.1
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
einops==0.6.0
filelock==3.11.0
fonttools==4.39.3
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.53.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sympy==1.11.1
systemd-python==234
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
torch==2.0.0
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.15
wadllib==1.3.6
Werkzeug==2.2.3
zipp==1.0.0
